Modify windows platform code to support x86_64 compilation

+++ engine/source/platform/types.win32.h
@@ -29,7 +29,7 @@
 // size_t is needed to overload new
 // size_t tends to be OS and compiler specific and may need to 
 // be if/def'ed in the future
-typedef unsigned int  dsize_t;      
+typedef size_t  dsize_t;
 
 
 /** Platform dependent file date-time structure.  The defination of this structure
